+++ src/communication/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 1.2.0
 ++++++
 * Add communication chat command group in preview mode.
+* Add communication identity command group in preview mode.
 
 
 1.1.2